MIGENG-150 fixed depending classes after moved class VMWorkloadinventory to another package
